### 2020-07-29 #2534 @kthyng
Hmm, all of our bioinformatics/genetics editors are pretty overloaded with submissions to handle. Do any of you feel able to take on this submission? @fboehm, @majensen, @will-rowe, @csoneson, @mikldk, @lpantano?
