### 2020-07-30 bblodfon/emba#1 @mikldk
Install using remotes instead of devtools
Consider changing from `devtools::install_github("bblodfon/emba")` to `remotes::install_github("bblodfon/emba")` as `remotes` has fewer dependencies. Via [JOSS](https://github.com/openjournals/joss-reviews/issues/2534).
