### 2020-08-06 #2534 @bblodfon
Hi @mikldk,

Just a kindly remainder that there has been no response from the potential reviewers regarding this paper submission (maybe they are on vacations :)
Should we try to contact them again or try finding others?
