+++ paper/paper.bib
@@ -488,5 +488,5 @@ year = {2020}
   title = {emba package website},
   year = {2020},
   publisher = {GitHub Pages},
-  url = {https://druglogics.github.io/druglogics-doc/}
+  url = {https://bblodfon.github.io/emba/}
 }
