### 2020-08-19 #2534 @edifice1989
Sure! I am in!

Ming

On Wed, Aug 19, 2020 at 1:29 AM Mikkel Meyer Andersen <
notifications@github.com> wrote:

>
>
> @edifice1989 <https://github.com/edifice1989>, @jkanche
> <https://github.com/jkanche>: Would you be interested in reviewing this
> submission to The Journal of Open Source Software
> <https://joss.theoj.org/>? Reviews are open and based on a checklist. The
> reviewer guidelines are available here:
> https://joss.readthedocs.io/en/latest/reviewer_guidelines.html. If you
> have any questions or concerns please let me know.
>
>
>
>
