### 2020-08-19 #2583 @whedon
[REVIEW]: emba: R package for analysis and visualization of biomarkers in boolean model ensembles
**Submitting author:** @bblodfon (<a href="http://orcid.org/0000-0002-3609-8674">John Zobolas</a>)
**Repository:** <a href="https://github.com/bblodfon/emba" target ="_blank">https://github.com/bblodfon/emba</a>
**Version:** v0.1.7
