### 2020-08-24 be817cc
filter input to gitsbe files only specify minimum version of package dependencies update NEWS.md update to version 0.1.7
