+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: emba
 Type: Package
 Title: Ensemble Boolean Model Biomarker Analysis
-Version: 0.1.6
+Version: 0.1.7
 Authors@R: person("John", "Zobolas", role = c("aut", "cph", "cre"), 
   email = "bblodfon@gmail.com", comment = c(ORCID = "0000-0002-3609-8674"))
 Description: Analysis and visualization of an ensemble of boolean models for 
@@ -35,12 +35,12 @@ Imports:
     visNetwork (>= 2.0.9),
     Ckmeans.1d.dp (>= 4.2.2),
     usefun (>= 0.4.3),
-    readr,
-    dplyr,
-    tidyr,
-    tidyselect,
-    stringr,
-    tibble
+    readr (>= 1.3.0),
+    dplyr (>= 1.0.0),
+    tidyr (>= 1.1.0),
+    tidyselect (>= 1.0.0),
+    stringr (>= 1.4.0),
+    tibble (>= 3.0.0)
 Suggests: 
     testthat,
     knitr,
