### 2020-08-26 #2534 @edifice1989
Hi Mikkel,

    The invitation link has expired for me
https://github.com/openjournals/joss-reviews/invitations

    Could you please resend it?

Thanks,
Ming

On Wed, Aug 19, 2020 at 5:24 AM Mikkel Meyer Andersen <
notifications@github.com> wrote:

> @whedon <https://github.com/whedon> add @edifice1989
> <https://github.com/edifice1989> as reviewer
>
