### 2020-08-27 #2534 @edifice1989
Yes. I have. But I can't click the checklist for me.
When I click the invitation link, it says 'the link has expired'

On Thu, Aug 27, 2020 at 3:50 AM Mikkel Meyer Andersen <
notifications@github.com> wrote:

> @edifice1989 <https://github.com/edifice1989> Don't you have access over
> at the review in #2583
> <https://github.com/openjournals/joss-reviews/issues/2583>?
>
