### 2020-08-27 #2583 @edifice1989
yes, problem solved!

Thank you very much!

Ming

On Thu, Aug 27, 2020 at 11:06 AM Mikkel Meyer Andersen <
notifications@github.com> wrote:

> @edifice1989 <https://github.com/edifice1989> Are you now able to edit
> your checklist?
>
