### 2020-08-27 #2583 @edifice1989
There is an issue when loading the 'emba' package. I opened a separate issue in  submission's [repository ](https://github.com/bblodfon/emba/issues/2#issue-687461142) please resolve it
