### 2020-08-27 bblodfon/emba#2 @bblodfon
Hi @edifice1989, I believe this has to do with me not specifying the exact minimum version of `tidyselect` in the `DESCRIPTION` file. Can you quickly check the version you are using with `packageVersion("tidyselect")`? (if it's `0.2.5` or earlier then there is no `all_of` in there :)
