### 2020-08-29 bblodfon/emba#2 @bblodfon
This should be fixed now - can you try now installing the development version with `remotes::install_github("bblodfon/emba")`? It should force now to install and import the newer version of `tidyselect`.

There is also a problem since the last 8 hours or so with https://travis-ci.com/ so I cannot test this version with CI right now, but with `devtools::check_rhub()` I tested its installation in 3 major platoforms and the results were succesfull. For example: https://builder.r-hub.io/status/emba_0.1.7.tar.gz-8feddab76c6b4695953769e92bfa5728 (`tidyselect` is installed at `1.1.0` version in a machine that does not have this library at all). I guess what happened in your case is that you had it installed a long time ago in a previous version and since I didn't specify the minimum version I needed, the installation did not update it and thus it failed to load the function I needed from that package.
