### 2020-08-31 #2583 @bblodfon
Issue bblodfon/emba#2 is now resolved. Please use the development version of the package if you stumble upon this issue (install with `remotes::install_github("bblodfon/emba")`) and after the reviews are over I will have the newer version (`0.1.7`) with all relevant changes uploaded to CRAN.
