### 2020-09-16 #2583 @mikldk
@Athanasiamo, @neerajdhanraj: Would one of you be interested in reviewing this submission to [The Journal of Open Source Software](https://joss.theoj.org/)? Reviews are open and based on a checklist. The reviewer guidelines are available here: <https://joss.readthedocs.io/en/latest/reviewer_guidelines.html>. If you have any questions or concerns please let me know.
