### 2020-09-16 #2583 @neerajdhanraj
I would like to review it for JoSS.

Regards,

Neeraj

On Wed, 16 Sep, 2020, 11:30 PM Mikkel Meyer Andersen, <
notifications@github.com> wrote:

> @Athanasiamo <https://github.com/Athanasiamo>, @neerajdhanraj
> <https://github.com/neerajdhanraj>: Would one of you be interested in
> reviewing this submission to The Journal of Open Source Software
> <https://joss.theoj.org/>? Reviews are open and based on a checklist. The
> reviewer guidelines are available here:
> https://joss.readthedocs.io/en/latest/reviewer_guidelines.html. If you
> have any questions or concerns please let me know.
>
