### 2020-09-17 #2583 @neerajdhanraj
Dear @mikldk ,

I provided my reviews to this paper by ticking the respective boxes. Is it the completion of the review process from my side? Or is there any other procedure to proceed?

Thank you.
