### 2020-09-21 #2583 @mikldk
@edifice1989, @neerajdhanraj:  Can you confirm that you have finished the review and recommend that this paper is now published?

@bblodfon :

* Please make a tagged release and archive (e.g. with [Zenodo](https://zenodo.org/)) as described [here](<https://joss.readthedocs.io/en/latest/submitting.html#the-review-process>), and report the version number and archive DOI in this thread. Please verify that the archive deposit has the correct metadata (title and author list), or edit these if that is not the case.
