### 2020-09-21 #2583 @edifice1989
Yes. I confirm! I recommend that this paper is now published.

Best,
Ming

On Mon, Sep 21, 2020 at 8:10 AM Mikkel Meyer Andersen <
notifications@github.com> wrote:

>
>
> @edifice1989 <https://github.com/edifice1989>, @neerajdhanraj
> <https://github.com/neerajdhanraj>: Can you confirm that you have
> finished the review and recommend that this paper is now published?
>
>
> @bblodfon <https://github.com/bblodfon> :
>
>
>
>
>
>    - Please have a final read though of the paper, checking language etc.
>
>    - Have a final check of the proofs with @whedon generate pdf
>
>    - Please make a tagged release and archive (e.g. with Zenodo
>    <https://zenodo.org/>) as described here
>    <https://joss.readthedocs.io/en/latest/submitting.html#the-review-process>,
>    and report the version number and archive DOI in this thread. Please verify
>    that the archive deposit has the correct metadata (title and author list),
>    or edit these if that is not the case.
>
>
>
>
>
>
