### 2020-09-22 285c171
add latex-based 'C++' text change order on some references (if it makes any difference)
