+++ paper/paper.md
@@ -33,7 +33,7 @@ bibliography: paper.bib
 # Introduction
 
 Computational modeling of cellular systems has been one of the most powerful tools used to build interpretable knowledge of biological processes and help identify molecular mechanisms that drive diseases such as cancer [@Aldridge2006].
-In particular, the use of logical modeling has proven to be a substantially useful approach, since it allows the easy construction, simulation and analysis of predictive models, capable of providing a qualitative and insightful view on the extremely complex landscape of biological systems [@Abou-Jaoude2016; @Morris2010; @Wang2012].
+In particular, the use of logical modeling has proven to be a substantially useful approach, since it allows the easy construction, simulation and analysis of predictive models, capable of providing a qualitative and insightful view on the extremely complex landscape of biological systems [@Morris2010; @Wang2012; @Abou-Jaoude2016].
 Such mechanistic models, with the systematic integration of prior knowledge and experimental data, have been extensively used to better understand what drives deregulation of signal transduction, the outcome of which is the manifestation of diseases [@Traynard2017].
 Furthermore, their explanatory power has been used to provide insights into a drug’s mode of action, investigate the mechanisms of resistance to drugs [@Eduati2017] and suggest new therapeutic combination candidates, among others [@Flobak2015].
 
@@ -55,7 +55,7 @@ Such an aggregation of best-fit models (wisdom of the crowds) has been shown to 
 There is a plethora of software tools devoted to the qualitative modeling and analysis of biological networks.
 The Consortium for the development of Logical Models and Tools (CoLoMoTo) is a community effort which aims to standardize the representation of logical networks and provide a common repository of methods and tools to analyze these networks [@Naldi2015].
 Furthermore, to facilitate the access to several software logical modeling tools and enable reproducible computational workflows, the CoLoMoTo Interactive Notebook was introduced as a unified computational framework [@Naldi2018a].
-The incorporated tools are accessed via a common programming interface (though originally implemented in different programming languages e.g. Java, Python, C++ and R) and offer a collection of features like accessing online model repositories [@Helikar2012], model editing [@Naldi2018b], dynamical analysis (finding attractors, stochastic simulations, reachability properties, model-checking techniques) [@Naldi2018; @Stoll2017; @Pauleve2017; @Klarner2016; @Mussel2010] and model parameterization/optimization to fit perturbation signaling data [@Terfve2012; @Gjerga2020].
+The incorporated tools are accessed via a common programming interface (though originally implemented in different programming languages e.g. Java, Python, C\texttt{++} and R) and offer a collection of features like accessing online model repositories [@Helikar2012], model editing [@Naldi2018b], dynamical analysis (finding attractors, stochastic simulations, reachability properties, model-checking techniques) [@Mussel2010; @Klarner2016; @Stoll2017; @Pauleve2017; @Naldi2018] and model parameterization/optimization to fit perturbation signaling data [@Terfve2012; @Gjerga2020].
 Despite the diverse and multi-purpose logical modeling tools that exist, there is still a lack of data analysis-oriented software that assists with the discovery of predictive biomarkers in ensembles of parameterized boolean networks that have been subject to drug combination perturbations.
 
 The `emba` R package aims to fill that gap and provide a first implementation of such a novel software.
