### 2020-09-22 9feba2a
update NEWS.md
