### 2020-09-22 #2583 @bblodfon
Hi @mikldk,

- I read the paper one final time and made sure everything is proper
- I see that one of the reviewers is **mentioned twice** in the generated paper (as is here in the first post, I guess they are connected)
- Archived version number for the package is `v0.1.7` (see [respective release](https://github.com/bblodfon/emba/releases/tag/v0.1.7))
- Archived Zenodo DOI is `10.5281/zenodo.4043085` => [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.4043085.svg)](https://doi.org/10.5281/zenodo.4043085)
- I made all appropriate edits in the Zenodo archived record

@edifice1989, @neerajdhanraj Thanks so much for reviewing my package!
