### 2020-09-23 #2583 @mikldk
@arfon Can you help with removing the double mention of @edifice1989 as reviewer? I am not sure how to handle that properly?
