### 2020-09-24 #2583 @arfon
OK, this looks fixed now in the PDF proofs?
