### 2020-09-25 #2583 @arfon
@bblodfon - could you please merge this small PR before we accept the paper: https://github.com/bblodfon/emba/pull/3 ? This just tweaks the formatting of a couple of references for improved readability.
