### 2020-09-26 #2583 @bblodfon
Hi @arfon, thanks for reading proof-reading my paper! I merged the PR. Though I think the **previous version** on the following sentence is better suited:
...suggest new therapeutic combination candidates, among others (Flobak at al. 2015). (**previous**)
...suggest new therapeutic combination candidates, among others Flobak et al. (2015). (**after the PR**) The way its written now allows for a more flexible interpretation I think, which is not what I intended. Is it ok with you to change it back to the previous version? What do you think @mikldk?
