### 2020-09-26 #2583 @arfon
> The way its written now allows for a more flexible interpretation I think, which is not what I intended. Is it ok with you to change it back to the previous version?

That's fine. I think I mis-interpreted the logical flow of the sentence. Feel free to revert!
